requirements.txt
# -----------------------------------------------------------------------------
# Database
# -----------------------------------------------------------------------------
mysql-connector-python==8.3.0

# -----------------------------------------------------------------------------
# Data processing
# -----------------------------------------------------------------------------
pandas>=1.0.0
numpy>=1.18.0

# -----------------------------------------------------------------------------
# Template rendering (HTML)
# -----------------------------------------------------------------------------
jinja2==3.1.4

# -----------------------------------------------------------------------------
# PDF generation
# -----------------------------------------------------------------------------
weasyprint>=0.40

# -----------------------------------------------------------------------------
# Configuration
# -----------------------------------------------------------------------------
pyyaml==6.0.2
python-dotenv==1.0.1

# -----------------------------------------------------------------------------
# Logging (optional enhancement)
# -----------------------------------------------------------------------------
loguru==0.7.2

# -----------------------------------------------------------------------------
# Date/time utilities
# -----------------------------------------------------------------------------
python-dateutil==2.9.0.post0

# -----------------------------------------------------------------------------
# Testing
# -----------------------------------------------------------------------------
pytest==8.2.2
